Modify sort.h file
Add print_list.c file
Separate the functions
