+++ print_array.c
@@ -20,24 +20,3 @@ void print_array(const int *array, size_t size)
 	}
 	printf("\n");
 }
-/**
- * print_list - Prints a list of integers
- *
- * @list: The list to be printed
- */
-void print_list(const listint_t *list)
-{
-	int i;
-
-	i = 0;
-	while (list)
-	{
-		if (i > 0)
-			printf(", ");
-		printf("%d", list->n);
-		++i;
-		list = list->next;
-	}
-	printf("\n");
-}
-
