+++ sort.h
@@ -18,7 +18,10 @@ typedef struct listint_s
 } listint_t;
 void print_array(const int *array, size_t size);
 void print_list(const listint_t *list);
-
+void bubble_sort(int *array, size_t size);
+void insertion_sort_list(listint_t **list);
+void selection_sort(int *array, size_t size);
+void quick_sort(int *array, size_t size);
